requirements.txt
flask==3.0.0
python-dotenv==1.0.0
transformers==4.36.2
torch==2.1.2
requests==2.31.0
werkzeug==3.0.1
flask-sqlalchemy==3.1.1
librosa==0.10.1